+++ SamTFE/Sources/GameMP/Game.cpp
@@ -170,16 +170,92 @@ static FLOAT gam_fChatSoundVolume = 0.25f;
 BOOL map_bIsFirstEncounter = FALSE;
 BOOL _bUserBreakEnabled = FALSE;
 
+void _FixTexturesOnObelisk(CTFileName strLevelName)
+{
+  // for each entity in the world
+  {FOREACHINDYNAMICCONTAINER(_pNetwork->ga_World.wo_cenEntities, CEntity, iten) {
+    // if it is brush entity
+    if (iten->en_RenderType == CEntity::RT_BRUSH) {
+      // for each mip in its brush
+      FOREACHINLIST(CBrushMip, bm_lnInBrush, iten->en_pbrBrush->br_lhBrushMips, itbm) {
+        // for all sectors in this mip
+        FOREACHINDYNAMICARRAY(itbm->bm_abscSectors, CBrushSector, itbsc) {
+          // for all polygons in sector
+          FOREACHINSTATICARRAY(itbsc->bsc_abpoPolygons, CBrushPolygon, itbpo)
+          {
+            CTFileName strTextureName = itbpo->bpo_abptTextures[1].bpt_toTexture.GetName().FileName();
+            int _Obelisk02Light_found   = strncmp((const char *)strTextureName, (const char *) "Obelisk02Light", (size_t) 14 );
+            if (_Obelisk02Light_found == 0 ){
+              if ( strLevelName=="KarnakDemo" || strLevelName=="Intro" || strLevelName=="08_Suburbs"
+                || strLevelName=="13_Luxor" || strLevelName=="14_SacredYards") {
+                itbpo->bpo_abptTextures[1].s.bpt_colColor = (C_WHITE| 0x5F);
+              } else if ( strLevelName=="04_ValleyOfTheKings" || strLevelName=="11_AlleyOfSphinxes" || strLevelName=="12_Karnak"){
+                itbpo->bpo_abptTextures[1].s.bpt_colColor = (C_GRAY| 0x2F);
+              }
+            }
+          }
+        }
+      }
+    } //
+  }}
+}
+
+void _FixTexturesOnAlleyOfSphinxes(void)
+{
+  // for each entity in the world
+  {FOREACHINDYNAMICCONTAINER(_pNetwork->ga_World.wo_cenEntities, CEntity, iten) {
+    // if it is brush entity
+    if (iten->en_RenderType == CEntity::RT_BRUSH) {
+      // for each mip in its brush
+      FOREACHINLIST(CBrushMip, bm_lnInBrush, iten->en_pbrBrush->br_lhBrushMips, itbm) {
+        // for all sectors in this mip
+        FOREACHINDYNAMICARRAY(itbm->bm_abscSectors, CBrushSector, itbsc) {
+          // for all polygons in sector
+          FOREACHINSTATICARRAY(itbsc->bsc_abpoPolygons, CBrushPolygon, itbpo)
+          {
+            CTFileName strTextureName = itbpo->bpo_abptTextures[1].bpt_toTexture.GetName().FileName();
+            int _EyeOfRa_found = strncmp((const char *)strTextureName, (const char *) "EyeOfRa", (size_t) 7 );
+            int _Wall12_found  = strncmp((const char *)strTextureName, (const char *) "Wall12",  (size_t) 6 );
+            int _Wingy02_found = strncmp((const char *)strTextureName, (const char *) "Wingy02", (size_t) 7 );
+            if (_EyeOfRa_found == 0 || _Wall12_found == 0 || _Wingy02_found == 0){
+              itbpo->bpo_abptTextures[1].s.bpt_ubBlend  = BPT_BLEND_BLEND;
+              itbpo->bpo_abptTextures[1].s.bpt_colColor = C_GRAY|0x80;
+            }
+          }
+        }
+      }
+    } //
+  }}
+}
+
 // make sure that console doesn't show last lines if not playing in network
 void MaybeDiscardLastLines(void)
 {
+  // Fix Textures Stuff
+  CTString strLevelName = _pNetwork->ga_fnmWorld.FileName();
+  //CPrintF("[strLevelName ==: %s]\n", (const char*)strLevelName);
+  CTString strModName = _pShell->GetValue("sys_strModName");
+  //CPrintF("[strModName ==: %s]\n", (const char*)strModName);
+
+  // Fix Obelisk textures
+  if ( strModName=="" ) {
+    if ( strLevelName=="04_ValleyOfTheKings" || strLevelName=="11_AlleyOfSphinxes" || strLevelName=="12_Karnak" 
+      || strLevelName=="13_Luxor" || strLevelName=="KarnakDemo" || strLevelName=="Intro" 
+      || strLevelName=="08_Suburbs" || strLevelName=="14_SacredYards") {
+      _FixTexturesOnObelisk(strLevelName);
+    }
+  }
+  // Fix Alley Of Sphinxes textures
+  if ( strModName=="" && strLevelName=="11_AlleyOfSphinxes") {
+    _FixTexturesOnAlleyOfSphinxes();
+  }
+
   // if not in network
   if (!_pNetwork->IsNetworkEnabled()) {
     // don't show last lines on screen after exiting console
     CON_DiscardLastLineTimes();
   }
 }
-
 
 class CEnableUserBreak {
 public: